requirements.txt
requests~=2.32.3
paho-mqtt~=2.1.0
python-dotenv==1.1.0